requirements.txt
Flask
boto3
mistune
psycopg[binary]
gunicorn
aws-opentelemetry-distro
opentelemetry-instrumentation-psycopg
bedrock-agentcore
